Make infrastructure more type safety friendly
