+++ gdsclient/call_builder.py
@@ -1,11 +1,10 @@
 from typing import Any, NoReturn
 
-from .algo.algo_endpoints import AlgoEndpoints
-from .graph.graph_endpoints import GraphEndpoints
+from .indirect_endpoints import IndirectEndpoints
 from .query_runner.query_runner import QueryRunner
 
 
-class CallBuilder(AlgoEndpoints, GraphEndpoints):
+class CallBuilder(IndirectEndpoints):
     def __init__(self, query_runner: QueryRunner, namespace: str):
         super().__init__(query_runner, namespace)
         self._query_runner = query_runner
@@ -15,5 +14,5 @@ class CallBuilder(AlgoEndpoints, GraphEndpoints):
         namespace = f"{self._namespace}.{attr}"
         return CallBuilder(self._query_runner, namespace)
 
-    def __call__(self, *args: Any, **kwargs: Any) -> NoReturn:
+    def __call__(self, *_: Any, **__: Any) -> NoReturn:
         raise SyntaxError(f"There is no {self._namespace} to call")
