+++ gdsclient/direct_endpoints.py
@@ -1,0 +1,8 @@
+from gdsclient.query_runner.query_runner import QueryRunner
+
+from .graph.graph_endpoints import GraphEndpoints
+
+
+class DirectEndpoints(GraphEndpoints):
+    def __init__(self, query_runner: QueryRunner, namespace: str):
+        super().__init__(query_runner, namespace)
